requirements.txt
fastapi
uvicorn
python-dotenv
mistralai
numpy
faiss-cpu
requests
tqdm
gradio